Разбил код бота на логику и функции получения данных и вынес их в bot_data_layer.py. Переименовал tools.py в db.py, так как там вся логика подключения и запросов к бд и все запросы. Переписал логику подключения к бд на connection pool. В дальнейшем хочу избавиться от пробросов соединений и сделать singleton для pool
